requirements.txt
django==4.0.6
python-telegram-bot==13.4.1